continuing to extract out dispatch
